Adjusted signature of libxs_coprime2 (uint->size_t).

+++ include/libxs_math.h
@@ -68,7 +68,7 @@ LIBXS_API size_t libxs_lcm(size_t a, size_t b);
 LIBXS_API int libxs_primes_u32(unsigned int num, unsigned int num_factors_n32[]);
 
 /** Calculate co-prime number <= n/2 (except: libxs_coprime2(0|1) == 0). */
-LIBXS_API size_t libxs_coprime2(unsigned int n);
+LIBXS_API size_t libxs_coprime2(size_t n);
 
 /**
  * Minimizes the waste, if "a" can only be processed in multiples of "b".
